fix(postcmd): log the command that runs, and never report a size as zero
Found by replaying a real update: an ngPost 5.4.2 configuration and history
database, driven by the previous binary, then handed to this one.

- post commands were logged as their template rather than as the command
  actually run, so the substituted values, which are the whole point of that
  log line, were no longer visible. The old ngPost printed them.

- a post made before this version records no rar+par2 size, and the sheet
  wrote "0" for it. An index importing the file would take a zero byte post
  at face value; an unknown size now renders empty, like the par2 percentage
  already did, and the export still says out loud what is missing.

A test covers what the update really risks: a 5.4.2 configuration, with its
server block, its two NZB_POST_CMD lines and its separator, must survive the
whole-file rewrite that a GUI save performs, and parse back unchanged.

+++ src/PostingJob.cpp
@@ -1464,7 +1464,7 @@ PostInfoData PostingJob::postInfoData() const
     data.avgSpeed   = avgSpeed();
     data.appVersion = QString(APP_VERSION);
 
-    data.postSizeBytes    = _postSizeBytes;
+    data.postSizeBytes    = static_cast<qint64>(_postSizeBytes);
     data.legacySizeBytes  = _totalSize;
     data.par2Pct          = _options.describedPar2Pct();
     data.nbFiles          = _nbFiles;
